+++ fork/AlchemyStars/src/AlchemyStars.Engine/IAnimationExportEngine.cs
@@ -1,0 +1,8 @@
+namespace AlchemyStars.Engine;
+
+public interface IAnimationExportEngine
+{
+    EngineCapabilities Capabilities { get; }
+
+    AnimationExportResult Export(AnimationExportRequest request);
+}